feat: counterPage 추가 및 분리
